Add flooder run script

+++ QOS/main.sh
@@ -5,8 +5,8 @@ touch insight.txt
 sleep 2s
 sh ezInsights.sh &
 
-mkdir web
-ln -s insight.txt web/insight.txt
+mkdir -p web
+test -f web/insight.txt || ln -s insight.txt web/insight.txt
 cd web
 cat /etc/hostname > hostname.txt
 echo $(ifconfig br0-int | grep -oE 'HWaddr .*' | grep -oE ' .*' | sed 's/ //g') > mac.txt
